+++ cmd/gopische/main.go
@@ -6,6 +6,7 @@ import (
 	"os"
 
 	"github.com/mnbi/gopische"
+	//	"github.com/mnbi/gopische/lexer"
 )
 
 var (